Measurable card empty state
